Fix block example
`mirage` will now generate an example .xl file which contains the path
to the file image, and `xl` will take care of setting up loop devices
if necessary. Previously we had to do this by hand, which meant that
running the Xen example was more complicated than running the Unix example.

Now the unikernel config.ml is independent of whether the target build is
for Xen or Unix.

Fixes #114

Signed-off-by: David Scott <dave.scott@docker.com>